fixed some logical errors

+++ src/main/java/com/emis/EMIS/wrappers/requestDTOs/MarksDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class MarksDTO {
     private int subjectId;
     private int studentId;
-    private double mark;
+    private int mark;
 }
